cargo_fish_app/requirements.txt
flask
requests
flask-mail
python-dotenv
